+++ src/commands/index.js
@@ -69,6 +69,7 @@ export * from './setex';
 export * from './setnx';
 export * from './sismember';
 export * from './smembers';
+export * from './smove';
 export * from './spop';
 export * from './srandmember';
 export * from './srem';